+++ lib/exceptions/enodoexception.py
@@ -12,3 +12,11 @@ class EnodoInvalidArgumentException(Exception):
 
 class EnodoInvalidDatasetSize(Exception):
     pass
+
+
+class EnodoScheduleException(Exception):
+
+    def __init__(self, message, job_config_name=None):
+        # Call the base class constructor with the parameters it needs
+        super().__init__(message)
+        self.job_config_name = job_config_name