+++ lib/siridb/siridb.py
@@ -1,3 +1,4 @@
+import logging
 import re
 
 from siridb.connector.lib.exceptions import (AuthenticationError, InsertError,
@@ -13,8 +14,8 @@ async def query_series_datapoint_count(siridb_client, series_name):
             f'select count() from "{series_name}"')
     except (QueryError, InsertError, ServerError, PoolError,
             AuthenticationError, UserAuthError) as e:
-        print(e)
-        pass
+        msg = str(e) or type(e).__name__
+        logging.error(f"Problem with SiriDB: {msg}")
     else:
         counts = result.get(series_name, [])
         if counts:
@@ -29,8 +30,8 @@ async def query_time_unit(siridb_client):
             f'show time_precision')
     except (QueryError, InsertError, ServerError, PoolError,
             AuthenticationError, UserAuthError) as e:
-        print("Connection problem with SiriDB server")
-        pass
+        msg = str(e) or type(e).__name__
+        logging.error(f"Problem with SiriDB: {msg}")
     if result is None:
         return None
     return result["data"][0]["value"]
@@ -43,7 +44,8 @@ async def query_series_data(siridb_client, series_name, selector="*"):
             f'select {selector} from "{series_name}"')
     except (QueryError, InsertError, ServerError, PoolError,
             AuthenticationError, UserAuthError) as e:
-        print("Connection problem with SiriDB server")
+        msg = str(e) or type(e).__name__
+        logging.error(f"Problem with SiriDB: {msg}")
         pass
     return result
 
@@ -54,8 +56,8 @@ async def drop_series(siridb_client, series_name):
         result = await siridb_client.query(f'drop series {series_name}')
     except (QueryError, InsertError, ServerError, PoolError,
             AuthenticationError, UserAuthError) as e:
-        print("Connection problem with SiriDB server")
-        pass
+        msg = str(e) or type(e).__name__
+        logging.error(f"Problem with SiriDB: {msg}")
     return result
 
 
@@ -65,9 +67,8 @@ async def insert_points(siridb_client, series_name, points):
         await siridb_client.insert({series_name: points})
     except (QueryError, InsertError, ServerError, PoolError,
             AuthenticationError, UserAuthError) as e:
-        print("Connection problem with SiriDB server")
-        print(e)
-        pass
+        msg = str(e) or type(e).__name__
+        logging.error(f"Problem with SiriDB: {msg}")
     return result
 
 
@@ -78,8 +79,8 @@ async def query_group_expression_by_name(siridb_client, group_name):
             f'list groups where name == "{group_name}"')
     except (QueryError, InsertError, ServerError, PoolError,
             AuthenticationError, UserAuthError) as e:
-        print("Connection problem with SiriDB server")
-        pass
+        msg = str(e) or type(e).__name__
+        logging.error(f"Problem with SiriDB: {msg}")
     groups = result.get('groups')
     if groups is None or len(groups) < 1:
         return None
@@ -89,56 +90,60 @@ async def query_group_expression_by_name(siridb_client, group_name):
 
 async def query_all_series_results(siridb_client, series_name, selector="*"):
     result = None
+    name_escaped = re.escape(series_name).replace('/', r'\/')
     try:
         result = await siridb_client.query(
             f'select {selector} from '
-            f'/enodo_{re.escape(series_name)}_.*?$/')
+            f'/enodo_{name_escaped}_.*?$/')
     except (QueryError, InsertError, ServerError, PoolError,
             AuthenticationError, UserAuthError) as e:
-        print("Connection problem with SiriDB server")
-        pass
+        msg = str(e) or type(e).__name__
+        logging.error(f"Problem with SiriDB: {msg}")
     return result
 
 
 async def query_series_forecasts(siridb_client, series_name, selector="*",
                                  only_future=False):
     result = None
+    name_escaped = re.escape(series_name).replace('/', r'\/')
     after = ""
     if only_future:
         after = " after now"
     try:
         result = await siridb_client.query(
             f'select {selector} from '
-            f'/enodo_{re.escape(series_name)}_forecast_.*?$/{after}')
+            f'/enodo_{name_escaped}_forecast_.*?$/{after}')
     except (QueryError, InsertError, ServerError, PoolError,
             AuthenticationError, UserAuthError) as e:
-        print("Connection problem with SiriDB server")
-        pass
+        msg = str(e) or type(e).__name__
+        logging.error(f"Problem with SiriDB: {msg}")
     return result
 
 
 async def query_series_anomalies(siridb_client, series_name, selector="*"):
     result = None
+    name_escaped = re.escape(series_name).replace('/', r'\/')
     try:
         result = await siridb_client.query(
             f'select {selector} from '
-            f'/enodo_{re.escape(series_name)}_anomalies_.*?$/')
+            f'/enodo_{name_escaped}_anomalies_.*?$/')
     except (QueryError, InsertError, ServerError, PoolError,
             AuthenticationError, UserAuthError) as e:
-        print("Connection problem with SiriDB server")
-        pass
+        msg = str(e) or type(e).__name__
+        logging.error(f"Problem with SiriDB: {msg}")
     return result
 
 
 async def query_series_static_rules_hits(
         siridb_client, series_name, selector="*"):
     result = None
+    name_escaped = re.escape(series_name).replace('/', r'\/')
     try:
         result = await siridb_client.query(
             f'select {selector} from '
-            f'/enodo_{re.escape(series_name)}_static_rules_.*?$/')
+            f'/enodo_{name_escaped}_static_rules_.*?$/')
     except (QueryError, InsertError, ServerError, PoolError,
             AuthenticationError, UserAuthError) as e:
-        print("Connection problem with SiriDB server")
-        pass
+        msg = str(e) or type(e).__name__
+        logging.error(f"Problem with SiriDB: {msg}")
     return result